fix(sonar): dedupe editor utils and harden decoding paths
